ISSUE-234: Fix activity titles with a lot of uppercase chars
