+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -2,7 +2,7 @@
 name: Bug report
 about: Create a report to help us improve
 title: ''
-labels: bug, improvement
+labels: bug
 assignees: robiningelbrecht
 
 ---
