+++ src/Infrastructure/Twig/StringTwigExtension.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Infrastructure\Twig;
+
+final readonly class StringTwigExtension
+{
+    public static function doEllipses(string $string, int $maxLength): string
+    {
+        if (strlen($string) <= $maxLength) {
+            return $string;
+        }
+
+        return mb_substr($string, 0, $maxLength - 3, 'UTF-8').'...';
+    }
+
+    public static function doRepeat(string $char, int $times): string
+    {
+        return str_repeat($char, $times);
+    }
+
+    public static function doCountUpperCaseChars(string $string): int
+    {
+        return strlen($string) - strlen(preg_replace('/[A-Z]/', '', $string));
+    }
+}
