+++ templates/svg/badge.svg.twig
@@ -21,10 +21,11 @@
             {% for activity in activities|slice(0, 5) %}
                 {% set translateY = 12.5 + (loop.index0 * 25) %}
                 {% set convertedDistance = activity.getDistance()|convertMeasurement %}
+                {% set maxTitleLength = 40 - (activity.getName()|slice(0, 40)|countUppercaseChars // 7 * 3) %}
                 <g transform="translate(5, {{ translateY }})">
                     <text x="0" y="0" text-anchor="start">
                         <tspan x="0">{{ activity.getStartDate()|formatDate() }}</tspan>
-                        <tspan x="65">{{ activity.getName()|ellipses(40) }}</tspan>
+                        <tspan x="65">{{ activity.getName()|ellipses(maxTitleLength) }}</tspan>
                         <tspan x="305" text-anchor="middle">
                             {{ convertedDistance.toFloat()|formatNumber(0) }}<tspan font-size="10px">{{ renderUnitSymbol('distance') }}</tspan>
                         </tspan>
@@ -94,7 +95,7 @@
         </g>
         <g transform="translate(0, 212)" font-family="'Segoe UI', Ubuntu, 'Helvetica Neue', Sans-Serif" font-size="10px"
            font-weight="normal">
-            <text x="30" y="12.5" fill="#bbbbbb">generated by https://github.com/robiningelbrecht/strava-statistics
+            <text x="42" y="12.5" fill="#bbbbbb">generated by https://github.com/robiningelbrecht/strava-statistics
             </text>
         </g>
     </g>
